preparing cache

+++ buildspec/install.sh
@@ -9,25 +9,34 @@ aws s3 cp s3://hygo.secrets/hygofit.env "$ENV_FILE"
 chmod 400 /tmp/hygo.pem
 
 # Update go
-echo "Updating go"
-(
-	if [ -d install/go ]; then exit 0; fi
-	mkdir -p install/go
-	cd install/go
-	curl -LO https://go.dev/dl/go1.24.4.linux-amd64.tar.gz >/dev/null
-	rm -rf /usr/local/go && tar -C /usr/local -xzf go1.24.4.linux-amd64.tar.gz >/dev/null
-)
 export PATH=/usr/local/go/bin:$PATH
-echo "Go Version"
-go version
+if go version | grep -q 1.24.4; then
+	echo "Go is up to date"
+else
+	echo "Updating go"
+	(
+		if [ -d install/go ]; then exit 0; fi
+		mkdir -p install/go
+		cd install/go
+		curl -LO https://go.dev/dl/go1.24.4.linux-amd64.tar.gz >/dev/null
+		rm -rf /usr/local/go && tar -C /usr/local -xzf go1.24.4.linux-amd64.tar.gz >/dev/null
+	)
+	eval "$EXPORT_GO"
+	echo "Go Version"
+	go version
+fi
 
 # Update node
-echo "Updating node"
-(
-	curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.40.3/install.sh | bash >/dev/null
-	source "$HOME/.nvm/nvm.sh"
-	nvm install 22 >/dev/null
-)
-source "$HOME/.nvm/nvm.sh"
-echo "Node Version"
-node -v
+if node -v | grep -q "$NODE_VERSION"; then
+	echo "Node is up to date"
+else
+	echo "Updating node"
+	(
+		curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.40.3/install.sh | bash >/dev/null
+		eval "$SOURCE_NVM"
+		nvm install "NODE_VERSION" >/dev/null
+	)
+	eval "$SOURCE_NVM"
+	echo "Node Version"
+	node -v
+fi
